refactor (rs-flowershop-121) constants path
