+++ src/app/slices/catalog.ts
@@ -3,7 +3,7 @@ import { createSlice, createAsyncThunk, PayloadAction } from '@reduxjs/toolkit';
 import CatalogService from '../services/catalog.service';
 import { RootState } from '../store';
 
-import { ITEMS_PER_PAGE } from '@/constants/const';
+import { ITEMS_PER_PAGE } from '@/constants';
 import { Status, IPageQueryResult, Categories } from '@/types/types';
 
 export const fetchProducts = createAsyncThunk('catalog/allProducts', async (_, thunkAPI) => {
