+++ src/components/UI/SortingBlock/conts.ts
@@ -1,0 +1,27 @@
+export const sortOptions = [
+  {
+    name: 'Default',
+    value: '',
+    id: 'default_option',
+  },
+  {
+    name: 'Price (low to high)',
+    value: 'price asc',
+    id: 'price_asc',
+  },
+  {
+    name: 'Price (high to low)',
+    value: 'price desc',
+    id: 'price_desc',
+  },
+  {
+    name: 'Name (A to Z)',
+    value: 'name.en-US asc',
+    id: 'name_asc',
+  },
+  {
+    name: 'Name (Z to A)',
+    value: 'name.en-US desc',
+    id: 'name_desc',
+  },
+];
