+++ src/pages/ProfilePage/NewAddressBlock/index.tsx
@@ -22,7 +22,7 @@ import SimpleCheckbox from '@/components/UI/FormFields/SimpleCheckbox';
 import SimpleInput from '@/components/UI/FormFields/SimpleInput';
 import SimpleSelect from '@/components/UI/FormFields/SimpleSelect';
 import { Typography } from '@/components/UI/Typography';
-import { countries } from '@/constants/const';
+import { countries } from '@/constants';
 import { AddressAction, ApiResponse, ProfileForm, Status, User } from '@/types/types';
 
 import s from '../ProfilePage.module.scss';
