+++ src/pages/RegisterPage/index.tsx
@@ -21,7 +21,7 @@ import SimpleCheckbox from '@/components/UI/FormFields/SimpleCheckbox';
 import SimpleInput from '@/components/UI/FormFields/SimpleInput';
 import SimpleSelect from '@/components/UI/FormFields/SimpleSelect';
 import { Typography } from '@/components/UI/Typography';
-import { countries } from '@/constants/const';
+import { countries } from '@/constants';
 import { CustomerData } from '@/types/types';
 
 import s from './RegisterPage.module.scss';
